==> 0_0_download.ipynb <==
from pathlib import Path

import requests
import sidrapy

def download_data():
    data_filepath = Path("data/agua-esgoto-desagregado.zip")
    if not data_filepath.exists():
        data_filepath.parent.mkdir(exist_ok=True, parents=True)
        url_data = (
            "https://github.com"
            "/viridisdata"
            "/snis-rawdata"
            "/releases"
            "/download"
            "/v2022.12.b"
            "/agua-esgoto-desagregado.zip"
        )
        r = requests.get(url_data)
        with open(data_filepath, "wb") as f:
            f.write(r.content)
    return data_filepath


def download_shp():
    shp_filepath = Path("data/BR_Municipios_2021.zip")
    if not shp_filepath.exists():
        shp_filepath.parent.mkdir(exist_ok=True, parents=True)
        shp_url = (
            "https://geoftp.ibge.gov.br"
            "/organizacao_do_territorio"
            "/malhas_territoriais"
            "/malhas_municipais"
            "/municipio_2021"
            "/Brasil"
            "/BR"
            "/BR_Municipios_2021.zip"
        )
        r = requests.get(shp_url, timeout=300)
        with open(shp_filepath, "wb") as f:
            f.write(r.content)
    return shp_filepath


def download_ipca():
    ipca_filepath = Path("data/ipca.csv")
    if not ipca_filepath.exists():
        ipca_filepath.parent.mkdir(exist_ok=True, parents=True)
        ipca = sidrapy.get_table(
            table_code="1737",
            territorial_level="1",
            ibge_territorial_code="1",
            period="all",
        )
        ipca.to_csv(ipca_filepath, encoding="utf-8", index=False)
    return ipca_filepath


def download_municipio():
    municipio_filepath = Path("data/municipio.csv.gz")
    if not municipio_filepath.exists():
        municipio_filepath.parent.mkdir(exist_ok=True, parents=True)
        municipio_url = (
            "https://storage.googleapis.com/basedosdados-public"
            "/one-click-download"
            "/br_bd_diretorios_brasil"
            "/municipio"
            "/municipio.csv.gz"
        )
        r = requests.get(municipio_url)
        with open(municipio_filepath, "wb") as f:
            f.write(r.content)
    return municipio_filepath

download_data()
download_shp()
download_ipca()
download_municipio()
