==> 1_1_populacao_atendida.ipynb <==
import re

import numpy as np
import pandas as pd

def read_raw_data(data_filepath):
    df = pd.read_csv(
        data_filepath,
        compression="zip",
        sep=";",
        decimal=",",
        thousands=".",
    )

    m = re.compile(r"^([\dA-Z_]+) - (.*)")

    def get_variable_code(x):
        mm = m.match(x)
        if mm:
            return mm.group(1)
        return x

    df = df.rename(columns=get_variable_code)

    return df

data_filepath = "data/agua-esgoto-desagregado.zip"
municipio_filepath = "data/municipio.csv.gz"

# READ
populacao_municipio_atendida = (
    read_raw_data(data_filepath)
    [
        [
            "ano_referencia",
            "POP_TOT",
            "AG001",
            "ES001",
        ]
    ]
    .groupby(["ano_referencia"])
    .sum()
    .reset_index()
)
populacao_municipio_atendida

populacao_municipio_atendida_2021 = (
    populacao_municipio_atendida.query("ano_referencia == 2021")
)
populacao_municipio_atendida_2021

populacao_total = populacao_municipio_atendida_2021["POP_TOT"].values[0]
populacao_com_agua = populacao_municipio_atendida_2021["AG001"].values[0]
populacao_com_esgoto = populacao_municipio_atendida_2021["ES001"].values[0]

populacao_sem_agua = populacao_total - populacao_com_agua
print(f"População sem água: {populacao_sem_agua / 1000000:.2f} mi {populacao_sem_agua / populacao_total * 100:.2f}%")

populacao_sem_esgoto = populacao_total - populacao_com_esgoto
print(f"População sem esgoto: {populacao_sem_esgoto / 1000000:.2f} mi {populacao_sem_esgoto / 210168717 * 100:.2f}%")
