==> 2_0_make_boxplot.ipynb <==
from pathlib import Path

import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns

def boxplot_tmr_uf(df_tmr_mun, year: int, ylim: tuple[float], plots_dir):
    # FIG - Violinplot da Tarifa Média por Grande Região do Brasil (2021)
    f, ax = plt.subplots(figsize=(20, 10))
    df_tmr_mun_year = df_tmr_mun.query(f"ano_referencia == {year}")
    uf_median = (
        df_tmr_mun_year
        .groupby("sigla_uf")
        [["tarifa_media"]]
        .median()
        .sort_values(by="tarifa_media")
    )
    boxplot_order_median = uf_median.index
    sns.boxplot(
        data=df_tmr_mun_year,
        ax=ax,
        x="sigla_uf",
        y="tarifa_media",
        order=boxplot_order_median,
        palette="viridis",
    )
    sns.despine()
    ax.set_xlabel("Região")
    ax.yaxis.set_major_formatter(lambda x, pos: f"{x:.02f}".replace(".", ","))
    ax.set_ylabel("Tarifa Média (R$)")
    ax.set_ylim(*ylim)
    f.tight_layout()
    f.savefig(plots_dir / "fig-tmr-uf-boxplot.png", dpi=300)
    plt.clf()

plots_dir = Path("plots")
plots_dir.mkdir(parents=True, exist_ok=True)
data_dir = Path("data")

df_tmr_mun = pd.read_csv(data_dir / "processed_tmr_mun.csv")

# CONFIGURATION
plt.rcParams["font.family"] = "serif"
plt.rcParams["font.size"] = 16

boxplot_tmr_uf(df_tmr_mun=df_tmr_mun, year=2021, ylim=(-1, 16), plots_dir=plots_dir)
