==> 2_2_make_timeseries_plots.ipynb <==
from pathlib import Path

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

def tsplot_tmr(df_regiao: pd.DataFrame, plots_dir: Path, nome_regiao: str, color: str):
    f, ax = plt.subplots(figsize=(20, 10))
    n_series = df_regiao["id_municipio"].nunique()
    alpha = 1 / (20 + np.log10(n_series))
    for id_municipio in df_regiao["id_municipio"].unique():
        df_regiao_mun = df_regiao.query(f"id_municipio == {id_municipio}")
        ax.plot(
            df_regiao_mun.ano_referencia,
            df_regiao_mun.tarifa_media_trend,
            color=color,
            alpha=alpha,
        )

    sns.despine(ax=ax)
    ax.set_xlabel("Ano de Referência")
    ax.xaxis.set_major_formatter(lambda x, pos: str(int(x)))
    ax.yaxis.set_major_formatter(lambda x, pos: f"{x:.02f}".replace(".", ","))
    ax.set_ylabel("Tarifa Média (R$)")
    ax.set_ylim(-1, 15)
    ax.set_title(f"Região: {nome_regiao}")
    f.tight_layout()
    f.savefig(plots_dir / f"fig-tmr-mun-trends-{nome_regiao}.png", dpi=300)
    plt.clf()
    plt.close()


def tsplot_tmr_gr(df_tmr_mun3, plots_dir):
    # FIG - Tendências das Tarifas Médias Municipais (1995-2021)
    regioes = {
        "Sul": "dodgerblue",
        "Sudeste": "green",
        "Centro-Oeste": "orange",
        "Nordeste": "red",
        "Norte": "purple",
    }
    for nome_regiao in regioes:
        print(nome_regiao)
        color = regioes[nome_regiao]
        df_regiao = df_tmr_mun3.query(f"nome_regiao == '{nome_regiao}'")

        tsplot_tmr(
            df_regiao=df_regiao,
            plots_dir=plots_dir,
            nome_regiao=nome_regiao,
            color=color,
        )

plots_dir = Path("img")
plots_dir.mkdir(parents=True, exist_ok=True)
data_dir = Path("data")

year_start = 2002

df_tmr_mun3 = pd.read_csv(data_dir / "processed_tmr_mun3.csv").query(
    f"ano_referencia >= {year_start}"
)

# CONFIGURATION
plt.rcParams["font.family"] = "serif"
plt.rcParams["font.size"] = 16

tsplot_tmr(
    df_regiao=df_tmr_mun3,
    plots_dir=plots_dir,
    nome_regiao="Brasil",
    color="black",
)
tsplot_tmr_gr(df_tmr_mun3=df_tmr_mun3, plots_dir=plots_dir)
